♻️ optimized structure
